% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridDataWrapper.R
\name{gridDataWrapper}
\alias{gridDataWrapper}
\title{Wrap data into a grid object}
\usage{
gridDataWrapper(data, copy_from_grid = NULL, dimNames = NULL, varname, lat,
  lon, timeVector = NULL, timeStart, intervalLength)
}
\description{
Wrap an array in a grid object (as defined in the `loadeR` bundle)
}
\author{
M. De Felice
}
